refactor: moving logic into separate package
