+++ cmd/auth_delete_key.go
@@ -1,11 +1,7 @@
 package cmd
 
 import (
-	"fmt"
-	"os"
-
-	"github.com/deta/deta-cli/api"
-	"github.com/deta/deta-cli/runtime"
+	"github.com/deta/deta-cli/cmd/logic"
 	"github.com/spf13/cobra"
 )
 
@@ -27,41 +23,7 @@ func init() {
 }
 
 func deleteAPIKey(cmd *cobra.Command, args []string) error {
-	wd, err := os.Getwd()
-	if err != nil {
-		return err
-	}
-	if len(args) != 0 {
-		wd = args[0]
-	}
-	runtimeManager, err := runtime.NewManager(&wd, false)
-	if err != nil {
-		return nil
-	}
-
-	isInitialized, err := runtimeManager.IsInitialized()
-	if err != nil {
-		return err
-	}
-
-	if !isInitialized {
-		return fmt.Errorf("no deta micro initialized in '%s'", wd)
-	}
-
-	progInfo, err := runtimeManager.GetProgInfo()
-	if err != nil {
-		return err
-	}
-
-	err = client.DeleteAPIKey(&api.DeleteAPIKeyRequest{
-		ProgramID: progInfo.ID,
-		Name:      apiKeyName,
-	})
-	if err != nil {
-		return err
-	}
-	fmt.Printf("Succesfully deleted api key '%s'\n", apiKeyName)
-	return nil
+	return logic.DeleteAPIKey(client, apiKeyName, args)
 }
 
 func authDeleteKeyExamples() string {
