+++ cmd/details.go
@@ -1,11 +1,7 @@
 package cmd
 
 import (
-	"fmt"
-	"os"
-
-	"github.com/deta/deta-cli/api"
-	"github.com/deta/deta-cli/runtime"
+	"github.com/deta/deta-cli/cmd/logic"
 	"github.com/spf13/cobra"
 )
 
@@ -23,50 +19,5 @@ func init() {
 }
 
 func details(cmd *cobra.Command, args []string) error {
-	wd, err := os.Getwd()
-	if err != nil {
-		return err
-	}
-	if len(args) != 0 {
-		wd = args[0]
-	}
-	runtimeManager, err := runtime.NewManager(&wd, false)
-	if err != nil {
-		return err
-	}
-	isInitialzied, err := runtimeManager.IsInitialized()
-	if err != nil {
-		return err
-	}
-	if !isInitialzied {
-		return fmt.Errorf("no deta micro initialized in '%s'", wd)
-	}
-	progInfo, err := runtimeManager.GetProgInfo()
-	if err != nil {
-		return err
-	}
-	if progInfo == nil {
-		return fmt.Errorf("failed to get deta micro details")
-	}
-	u, err := getUserInfo(runtimeManager, client)
-	if err != nil {
-		return err
-	}
-	res, err := client.GetProjects(&api.GetProjectsRequest{
-		SpaceID: u.DefaultSpace,
-	})
-	if err != nil {
-		return err
-	}
-	for _, p := range res.Projects {
-		if p.ID == progInfo.Project {
-			progInfo.Project = p.Name
-		}
-	}
-	output, err := progInfoToOutput(progInfo)
-	if err != nil {
-		return err
-	}
-	fmt.Println(output)
-	return nil
+	return logic.Details(client, args)
 }
