+++ cmd/logic/auth_delete_key.go
@@ -1,0 +1,47 @@
+package logic
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/deta/deta-cli/api"
+	"github.com/deta/deta-cli/runtime"
+)
+
+func DeleteAPIKey(client *api.DetaClient, apiKeyName string, args []string) error {
+	wd, err := os.Getwd()
+	if err != nil {
+		return err
+	}
+	if len(args) != 0 {
+		wd = args[0]
+	}
+	runtimeManager, err := runtime.NewManager(&wd, false)
+	if err != nil {
+		return nil
+	}
+
+	isInitialized, err := runtimeManager.IsInitialized()
+	if err != nil {
+		return err
+	}
+
+	if !isInitialized {
+		return fmt.Errorf("no deta micro initialized in '%s'", wd)
+	}
+
+	progInfo, err := runtimeManager.GetProgInfo()
+	if err != nil {
+		return err
+	}
+
+	err = client.DeleteAPIKey(&api.DeleteAPIKeyRequest{
+		ProgramID: progInfo.ID,
+		Name:      apiKeyName,
+	})
+	if err != nil {
+		return err
+	}
+	fmt.Printf("Succesfully deleted api key '%s'\n", apiKeyName)
+	return nil
+}
