+++ cmd/logic/run_test.go
@@ -1,4 +1,4 @@
-package cmd
+package logic
 
 import (
 	"reflect"
