+++ cmd/logic/utils.go
@@ -1,4 +1,4 @@
-package cmd
+package logic
 
 import (
 	"encoding/json"
